RecordSpanExtensions
