+++ Mutagen.Bethesda.Core.UnitTests/Plugins/Binary/UtilityTranslationTests.cs
@@ -48,13 +48,13 @@ public class UtilityTranslationTests
     public void EnumerateSubrecords_Empty()
     {
         byte[] b = new byte[0];
-        Assert.Empty(PluginUtilityTranslation.EnumerateSubrecords(new ReadOnlyMemorySlice<byte>(b), GameConstants.Oblivion));
+        Assert.Empty(RecordSpanExtensions.EnumerateSubrecords(new ReadOnlyMemorySlice<byte>(b), GameConstants.Oblivion));
     }
 
     [Fact]
     public void EnumerateSubrecords_Typical()
     {
-        var ret = PluginUtilityTranslation.EnumerateSubrecords(GetTypical(), GameConstants.Oblivion).ToArray();
+        var ret = RecordSpanExtensions.EnumerateSubrecords(GetTypical(), GameConstants.Oblivion).ToArray();
         Assert.Equal(2, ret.Length);
         Assert.Equal(new RecordType("EDID"), ret[0].Key);
         Assert.Equal(FirstTypicalLocation, ret[0].Value);
@@ -65,7 +65,7 @@ public class UtilityTranslationTests
     [Fact]
     public void EnumerateSubrecords_Duplicate()
     {
-        var ret = PluginUtilityTranslation.EnumerateSubrecords(GetDuplicate(), GameConstants.Oblivion).ToArray();
+        var ret = RecordSpanExtensions.EnumerateSubrecords(GetDuplicate(), GameConstants.Oblivion).ToArray();
         Assert.Equal(3, ret.Length);
         Assert.Equal(new RecordType("EDID"), ret[0].Key);
         Assert.Equal(FirstTypicalLocation, ret[0].Value);
@@ -81,7 +81,7 @@ public class UtilityTranslationTests
     public void FindFirstSubrecords_Empty()
     {
         var b = new byte[0];
-        var ret = PluginUtilityTranslation.FindFirstSubrecords(b, GameConstants.Oblivion, FirstTypicalType, SecondTypicalType);
+        var ret = RecordSpanExtensions.FindFirstSubrecords(b, GameConstants.Oblivion, FirstTypicalType, SecondTypicalType);
         Assert.Equal(2, ret.Length);
         Assert.Null(ret[0]);
         Assert.Null(ret[1]);
@@ -90,7 +90,7 @@ public class UtilityTranslationTests
     [Fact]
     public void FindFirstSubrecords_Typical()
     {
-        var ret = PluginUtilityTranslation.FindFirstSubrecords(GetTypical(), GameConstants.Oblivion, SecondTypicalType, FirstTypicalType);
+        var ret = RecordSpanExtensions.FindFirstSubrecords(GetTypical(), GameConstants.Oblivion, SecondTypicalType, FirstTypicalType);
         Assert.Equal(2, ret.Length);
         Assert.Equal(SecondTypicalLocation, ret[0]?.Location);
         Assert.Equal(FirstTypicalLocation, ret[1]?.Location);
@@ -99,7 +99,7 @@ public class UtilityTranslationTests
     [Fact]
     public void FindFirstSubrecords_Single()
     {
-        var ret = PluginUtilityTranslation.FindFirstSubrecords(GetTypical(), GameConstants.Oblivion, SecondTypicalType);
+        var ret = RecordSpanExtensions.FindFirstSubrecords(GetTypical(), GameConstants.Oblivion, SecondTypicalType);
         Assert.Single(ret);
         Assert.Equal(SecondTypicalLocation, ret[0]?.Location);
     }
@@ -107,7 +107,7 @@ public class UtilityTranslationTests
     [Fact]
     public void FindFirstSubrecords_Duplicate()
     {
-        var ret = PluginUtilityTranslation.FindFirstSubrecords(GetDuplicate(), GameConstants.Oblivion, SecondTypicalType, FirstTypicalType);
+        var ret = RecordSpanExtensions.FindFirstSubrecords(GetDuplicate(), GameConstants.Oblivion, SecondTypicalType, FirstTypicalType);
         Assert.Equal(2, ret.Length);
         Assert.Equal(SecondTypicalLocation, ret[0]?.Location);
         Assert.Equal(FirstTypicalLocation, ret[1]?.Location);
@@ -119,19 +119,19 @@ public class UtilityTranslationTests
     public void FindFirstSubrecord_Empty()
     {
         var b = new byte[0];
-        Assert.Null(PluginUtilityTranslation.FindFirstSubrecord(b, GameConstants.Oblivion, SecondTypicalType));
+        Assert.Null(RecordSpanExtensions.FindFirstSubrecord(b, GameConstants.Oblivion, SecondTypicalType));
     }
 
     [Fact]
     public void FindFirstSubrecord_Typical()
     {
-        Assert.Equal(SecondTypicalLocation, PluginUtilityTranslation.FindFirstSubrecord(GetTypical(), GameConstants.Oblivion, SecondTypicalType)?.Location);
+        Assert.Equal(SecondTypicalLocation, RecordSpanExtensions.FindFirstSubrecord(GetTypical(), GameConstants.Oblivion, SecondTypicalType)?.Location);
     }
 
     [Fact]
     public void FindFirstSubrecord_Duplicate()
     {
-        Assert.Equal(FirstTypicalLocation, PluginUtilityTranslation.FindFirstSubrecord(GetTypical(), GameConstants.Oblivion, FirstTypicalType)?.Location);
+        Assert.Equal(FirstTypicalLocation, RecordSpanExtensions.FindFirstSubrecord(GetTypical(), GameConstants.Oblivion, FirstTypicalType)?.Location);
     }
     #endregion
 }
