+++ Mutagen.Bethesda.Core/Mutagen.Bethesda.Core.csproj
@@ -183,6 +183,7 @@
       <BuildAction>Compile</BuildAction>
     </Compile>
     <Compile Include="Plugins\Binary\Translations\PreviousParse.cs" />
+    <Compile Include="Plugins\Binary\Translations\RecordSpanExtensions.cs" />
     <Compile Include="Plugins\Binary\Translations\TypedParseParams.cs" />
     <Compile Include="Plugins\Binary\Translations\TypedWriteParams.cs" />
     <Compile Include="Plugins\Cache\IIdentifierLinkCache.cs" />
